+++ examples/resources/fmc_ipv6_prefix_lists/resource.tf
@@ -4,7 +4,7 @@ resource "fmc_ipv6_prefix_lists" "example" {
       entries = [
         {
           action            = "PERMIT"
-          ip_address        = "2001:1::/64"
+          prefix            = "2001:1::/64"
           min_prefix_length = 65
           max_prefix_length = 120
         }