+++ internal/provider/data_source_fmc_ipv6_prefix_list.go
@@ -89,7 +89,7 @@ func (d *IPv6PrefixListDataSource) Schema(ctx context.Context, req datasource.Sc
 							MarkdownDescription: "Action to take.",
 							Computed:            true,
 						},
-						"ip_address": schema.StringAttribute{
+						"prefix": schema.StringAttribute{
 							MarkdownDescription: "IPv6 address with prefix length.",
 							Computed:            true,
 						},