+++ internal/provider/data_source_fmc_ipv6_prefix_lists_test.go
@@ -33,7 +33,7 @@ func TestAccDataSourceFmcIPv6PrefixLists(t *testing.T) {
 	checks = append(checks, resource.TestCheckResourceAttrSet("data.fmc_ipv6_prefix_lists.test", "items.my_ipv6_prefix_lists.id"))
 	checks = append(checks, resource.TestCheckResourceAttrSet("data.fmc_ipv6_prefix_lists.test", "items.my_ipv6_prefix_lists.type"))
 	checks = append(checks, resource.TestCheckResourceAttr("fmc_ipv6_prefix_lists.test", "items.my_ipv6_prefix_lists.entries.0.action", "PERMIT"))
-	checks = append(checks, resource.TestCheckResourceAttr("fmc_ipv6_prefix_lists.test", "items.my_ipv6_prefix_lists.entries.0.ip_address", "2001:1::/64"))
+	checks = append(checks, resource.TestCheckResourceAttr("fmc_ipv6_prefix_lists.test", "items.my_ipv6_prefix_lists.entries.0.prefix", "2001:1::/64"))
 	checks = append(checks, resource.TestCheckResourceAttr("fmc_ipv6_prefix_lists.test", "items.my_ipv6_prefix_lists.entries.0.min_prefix_length", "65"))
 	checks = append(checks, resource.TestCheckResourceAttr("fmc_ipv6_prefix_lists.test", "items.my_ipv6_prefix_lists.entries.0.max_prefix_length", "120"))
 	resource.Test(t, resource.TestCase{
@@ -61,7 +61,7 @@ func testAccDataSourceFmcIPv6PrefixListsConfig() string {
 	config += `	items = { "my_ipv6_prefix_lists" = {` + "\n"
 	config += `		entries = [{` + "\n"
 	config += `			action = "PERMIT"` + "\n"
-	config += `			ip_address = "2001:1::/64"` + "\n"
+	config += `			prefix = "2001:1::/64"` + "\n"
 	config += `			min_prefix_length = 65` + "\n"
 	config += `			max_prefix_length = 120` + "\n"
 	config += `		}]` + "\n"